frontend/requirements_frontend.txt
streamlit==1.50.0
requests==2.32.5   
matplotlib==3.10.6
altair==5.5.0
pydeck==0.9.1
plotly         
pillow==11.3.0
opencv-python==4.12.0.88   